Locking the modal events when async operations are in progress
